fix: avoid autocomplete popover to be too low
Note: new InputControl container `.mud-input-control-input-container` enables relative positioning so that absolute positioning of popover works as expected.
Closes #472
fix: re-enable label to animate in autocomplete

Since the Autocomplete Popover was moved between `.mud-shrink` or `.mud-input ` & the `label`, changed the selector for a more tolerant one to apply reduced style on label *after* the `.mud-shrink` or `.mud-input` (instead of *strictly just after*).