DatePicker: Fix setting culture not leading to DateFormat change (#2962)
MudBaseDatePicker sets the Culture and Format on the default converter.
The CurrentCulture is used, and the CurrentCulture ShortDateFormat is used for the Format.

When DateFormat is set the converters Format is replaced.

When DateForamt is not set but Culture is set on the component this led to the unexpected behavior
of the CurrentCulture date format being used instead of the culture that was set (as reported in #2962).

With this change, setting the culture also leads to a Format change unless the DateFormat property has been set.

The Terminology “touched” is used for the private field to match the Touched property terminology.

Fixes #2962