fix #3847: MudSelect scroll to selected item, when list is long
We have the element reference, so a position calculation is not required.
We just set the containers scroll position so the element is in view.

The MudAutocomplete had to be fixed as well.

The SelectUsageExample had added some more choices and a default value, to demonstrate
the working of the patch.