close #4761 MudDiaog: CloseOnEscapeKey handling
This fix was more complex than intended.

The reason was, that the IKeyInterceptor was instanciated for all MudDialogInstance and thus also captured events for all instances.

Instead of disabling/enabling the IKeyInterceptor for each instance the IKeyInterceptor was moved to the MudDialogProvider and the KeyDown
event is forwarded to the top level dialog only.

This required to capture the MudDialogInstance and store it within IDialogReference. I don't like the idea of extending the interface too much, but
I saw no other option.

Also I introduced a new interface IDialogKeyboardHandler which allows the top level dialog to handle any other future keyboard events. I did not incoperate any calls ot
IKeyInterceptor.UpdateKey(options) yet, since there was no need yet and I wanted to keep the change as small as possible.

I'm quite open for discussions, but this bug #4761 needs a fix, because it closes the wrong dialog instance and even more bad it reverted all Parameters to their default!