close #4871 MudDialog instance confusion
In my opinion this fix is quite important, since it cleans up an the handling of duplicate
dialog parameter setting within DialogService.

This could fix many problems, if many dialogs are open and an dialog is removed not in the
reverse order it was shown. e.g. scenarios where you have a next button of a dialog so.

It does this by providing the @key attribute when instantiating dialogs. This leaves
the hard work if the component needs to be reinitialized or the parameters changed to
the framework, instead of building an own way.