MudHighlighter: Reduce memory allocations
Cache StringBuilder.
Avoid LINQ  usage.
Replace IEnumerable<string> return type to Memory<string>. This allows to segment and reuse the array returned by Regex.Split method.
When no result is returned, instead of returning an empty List, returns an empty Memory, which doesn't allocate.
When a single result is returned, instead of returning it inside a List, return it inside an array, which saves one allocation (List allocates an array under the hood).
Note: additional things could be done to avoid more allocations, such using StringBuilder.EnsureCapacity(int), replace Regex.Escape(string) with a custom implementation that instead of returning an string it appends the result directly on the StringBuilder and checking the underlying type of the IEnumerable<string> highlightedTexts parameter to avoid allocating its enumerator.