This work is related to working on the git issue: https://github.com/MudBlazor/MudBlazor/issues/6181
It provides a pathway to set some KeyOptions up on a MudNumericField and if that list is not null, we can parse it to replace problematic characters by leveraging a new KeyOptions.Replace field to scrub the string prior to setting it as the new value.

I was unable to understand why the numpad . wouldn't work like the general . on the main part of the keyboard for the fr-FR culture, but I believe it to be baked into the CultureInfo behavior and not MudBlazor (unless I missed something).

I'm willing to take any and all feedback or suggestions for different ways to achieve this.

I did refactor some tests to generify them and simplify them, which is - I believe - the only rule I broke on the commit list, however, I preserved the original tests in doing so.  They are now just row tests instead.

I do wonder about making a KeyOptionsFactory to assist in building the KeyOptions list to replace keystrokes, I also wonder if maybe this should be put in the MudInput so it could be leveraged by other fields.

I originally looked to try and leverage the Javascript to achieve this with the KeyInterceptor, but I wasn't able to see how to get that pathway to fire off.

Let me know where to take this, if you want me to reimagine how I'm achieving the desired result.