Docs: Always show search bar
Every time I go to the homepage of the docs the search bar is an extra click away, but more importantly it's not there at all on mobile and you can only see an icon to go to the GitHub, but on the homepage right in front of you is a huge button to do the same thing.

- Remove condition and always show search bar
- Makes sure Notifications and GitHub buttons have tooltips.
- Reduce space between search and buttons to fit better on narrow screens that aren't quite mobile

I also had to copy the grey styles from the docs page to the landing page or the searchbar was a weird color. I think the landing page and docs pages should probably share themes but that's for another time.