ToggleGroup: Fix border rounding overflow on selected item
If you enabled the Rounded property it would be fine, but the inherited border-radius from MudTheme would overflow. This fixes that and simplifies the implementation by hiding the overflow instead of calculating the rounding.

Also reverts a fourth toggle button in the example added in #8377 which made the whole thing too cramped when CheckMarks were enabled.