Docs: Expand subgroups by default
Resolves #1671

I don't see the point in having these collapsed by default. They should be first class citizens with the rest of the components and the groups should just be for organization.

Ctrl+F has always been a pain because of this, and since the search doesn't have a hotkey, you have to use your mouse more than you should have to.