Docs: Revert landing page search bar to icon
Original design was preferred and since the auto-focus in the dialog was fixed, it's the same number of clicks to search anyway.

Ideally we want a keyboard shortcut for this in the future: https://github.com/MudBlazor/MudBlazor/pull/7268/#issuecomment-2054120627

Reverts part of #8515 and Closes #8596

This also reverts the change in #8682 to the size of the search dialog since now it's no longer exclusive to small screens.