Dialog: Stretch content so actions are at the end
Apparently this was a regression which I expect was caused by the addition of a Focus Trap. The dialog content was expecting a flex parent, but the focus trap intercepted that.

This adds classes to the MudFocusTrap and the container for the child content so you can apply styles and debug the layout easier by default.

With those extra classes I applied flex to the intemediary components to make the dialog content grow as intended.