MudText: Ability to control which HTML tag is used
Resolves #8898 by adding HtmlTag.

Resolves #5535 by removing the `Inline` tag and making consumers use our more flexible and explicit CSS utility classes: https://mudblazor.com/utilities/display. E.g. `<MudText Typo="Typo.caption" Class="d-inline">`. This simplifies the library by removing an underused property.

WARNING: If the removal of `Inline` is accepted I still need to go back and adjust the few texts in the docs that use it.

Adds comprehrensive documentation to the component and plenty to the docs.

Adds comprehensive tests as there were none before.

I could have used `MudElement` but that seemed too DRY. This gives us more flexibility and keeps the code simple as it is a fundamental component on the same level. Spaghetti be gone.