Make object-fit and object-position styles `important`
This adds the `!important` override to the `object-fit` and `object-position` utility classes, much like other utility classes. I suspect those were added later and these two were left behind or forgotten.

Technically a breaking change but I don't think many people if any are using this utility and I believe it's only used in the MudAvatar in MudBlazor. This should be the expected behavior anyway.

Closes #9218