Deprecate several Style properties
Styling the component from inside the markup should never be preferred to doing it with classes that corrospond to CSS in stylesheets.

We currently avoid adding style properties but have some left over. This starts us on the path of losing them entirely.

DataGrid and Table were left alone as it's more complex and harder to change for both us and the developer. MudTimelineItem.DotStyle did not have a corresponding class so that was left too.