Removed hasfocusedasync, no longer needed to stop jumping down on second view of menu Formatted cascading value to match rest of page Final arrow left, arrow right and tabs changed to return to menu overview button as seen here https://www.w3.org/WAI/ARIA/apg/patterns/menubar/examples/menubar-navigation/
Bugs: Custom activator content not tabbing back to button on escape or on tab as the ref would need to be within the activator content or with custom js. If you could let me know how you would like me to deal with this.
Two-way Binding button doesnt return to the button when closed, the tabbing returns to the end of the popover